Converted coordinate system to positve/negative intengers.
